feat: 资源下负载均衡新增同步功能 --story=120475847
